console: Refactor application overview to extract connect function
